Use async inventory for graph refreship
When doing full graph refresh use the aync batching to get the inventory from the provider.
This addresses: https://bugzilla.redhat.com/show_bug.cgi?id=1507773
