+++ app/models/manageiq/providers/redhat/inventory_collection_default/infra_manager.rb
@@ -13,6 +13,7 @@ class ManageIQ::Providers::Redhat::InventoryCollectionDefault::InfraManager < Ma
           :name,
           :location,
           :template,
+          :memory_limit,
           :memory_reserve,
           :raw_power_state,
           :boot_time,
@@ -40,6 +41,7 @@ class ManageIQ::Providers::Redhat::InventoryCollectionDefault::InfraManager < Ma
           :name,
           :location,
           :template,
+          :memory_limit,
           :memory_reserve,
           :raw_power_state,
           :boot_time,
